+++ tests/SmartMT/test_startGUI.py
@@ -10,9 +10,9 @@ from qtpy.QtTest import QTest
 from mtpy.gui.SmartMT.start import StartGUI
 from tests import EDI_DATA_DIR
 from tests.SmartMT import _click_area
-from tests.imaging import reset_matplotlib
 
 _pos_check_box = QPoint(8, 8)
+
 
 @pytest.mark.last
 class TestStartGUI(TestCase):
@@ -22,7 +22,6 @@ class TestStartGUI(TestCase):
     """
     @classmethod
     def setUpClass(cls):
-        reset_matplotlib()
         import matplotlib.pyplot as plt
         plt.interactive(False)
 
@@ -39,7 +38,6 @@ class TestStartGUI(TestCase):
     def test_plot_mt_response_default(self):
         self._switch_to_plot("MT Response")
         self._plot()
-        self.assertTrue(self.smartMT._subwindow_counter == 1, "no image created")  # test if the image is created
 
     def test_plot_mt_response_enable_all(self):
         self._switch_to_plot("MT Response")
@@ -54,6 +52,10 @@ class TestStartGUI(TestCase):
 
         self._plot()
         self.assertTrue(self.smartMT._subwindow_counter == 1, "no image creataed")  # test if the image is created
+
+    def test_multiple_mt_response(self):
+        self._switch_to_plot("Multiple MT responses")
+        self._plot()
 
     def _switch_to_plot(self, name):
         # load some data
@@ -81,10 +83,14 @@ class TestStartGUI(TestCase):
         self.smartMT._progress_bar.onFinished()
 
     def _plot(self, timeout=3000):
+        subwindow_counter = self.smartMT._subwindow_counter
+
         loop = QtCore.QEventLoop()
         self.smartMT._plot_option._current_plot.plotting_completed.connect(loop.quit)
         if timeout is not None:
             QtCore.QTimer.singleShot(timeout, loop.quit)
         _click_area(self.smartMT._plot_option.ui.pushButton_plot)  # wait for plotting
         loop.exec_()
+
+        self.assertTrue(self.smartMT._subwindow_counter == subwindow_counter + 1, "no image created")  # test if the image is created
